+++ helpers/core/__init__.py
@@ -1,6 +1,9 @@
 import pandas as pd
-from env import __version__
-from helpers import print_message, print_choices, print_choice, INFO, WARNING, ERROR, loader, SYSTEM_PLATFORM, get_system_based_value
+import glob
+import os
+from pathlib import Path
+from env import __version__, LOGS_PATH
+from helpers import print_message, print_choices, print_choice, render_code, INFO, WARNING, ERROR, loader, SYSTEM_PLATFORM, get_system_based_value
 from helpers.core.utils import install_update, retrieve_json
 from helpers.modules.utils import github_url_to_releases_api
 
@@ -20,18 +23,19 @@ __methods_static_aliases__ = {}
 
 class coreHelper:
     """
-    Module to work with helper core.
+    **Module to work with helper core.**
     """
     def __init__(self, settings: dict):
         self.settings = settings.get('core')
 
         # Subclasses
-        self.config = self.config(dict(loader.config), settings.get('core:config'))
-        self.update = self.update(settings.get('core:update'))
+        self.config = self.config(dict(loader.config), settings.get('core:config', {}))
+        self.update = self.update(settings.get('core:update', {}))
+        self.logs = self.logs(settings.get('core:logging', {}))
 
     class config:
         """
-        Module to manipulate helper config parameters.
+        **Module to manipulate helper config parameters.**
         """
         def __init__(self, global_settings: dict, settings: dict):
             self.global_settings = global_settings
@@ -39,20 +43,24 @@ class coreHelper:
 
         def get(self, section: str, option: str):
             """
-            Method retrieves config data.
+            **Method retrieves config data.**
+            ```
             Usage:
                 core config get <section> <option>
+            ```
             """
             return f"Config data is:\n  Section: {section}\n  Option: {option}\n  Value: {loader.get(section, option)}"
         
         def set(self, section: str, option: str, value: str, system_based: bool = False):
             """
-            Method sets config data.
+            **Method sets config data.**
+            ```
             Usage:
                 1. Simple value set:
                     core config set <section> <option> <value>
                 2. Value set based on current system:
                     core config set <section> <option> <value> --system-based
+            ```
             """
             if system_based:
                 current_value = loader.get(section, option)
@@ -71,9 +79,11 @@ class coreHelper:
 
         def ls(self, pretty: bool = True):
             """
-            Method displays all config data.
+            **Method displays all config data.**
+            ```
             Usage:
                 core config ls
+            ```
             """
             if pretty:
                 for section_name, section_proxy in self.global_settings.items():
@@ -86,7 +96,7 @@ class coreHelper:
 
     class update:
         """
-        Module updates helper core.
+        **Module updates helper core.**
         """
         def __init__(self, settings: dict):
             self.settings = settings
@@ -94,10 +104,13 @@ class coreHelper:
 
         def check(self, pretty: bool = True, dev: bool = False):
             """
-            Method checks updates for helper core.
+            **Method checks updates for helper core.**
+            ```
             Usage:
                 core update check
+            ```
             """
+            dev = self.settings.get('enable_dev_updates', 'False') == 'True' or dev
             if not self.core_url:
                 return [] if not pretty else "Failed to retrieve core updates."
 
@@ -117,11 +130,11 @@ class coreHelper:
                     core_remote_changelog = core_last_release.get('body')
                     break
 
-            if core_remote_version and (__version__ < core_remote_version or (__version__ <= core_remote_version and 'dev' in core_last_release['tag_name'])):
+            if core_remote_version and (__version__ < core_remote_version or (__version__ < core_remote_version and 'dev' in core_last_release['tag_name'])):
                 update_data = {
                     'name': 'core',
                     'current_version': __version__,
-                    'remote_version': core_remote_version,
+                    'remote_version': core_remote_version + ' (dev)' if is_dev_release else '',
                 }
                 if not pretty:
                     update_data.update({
@@ -134,15 +147,99 @@ class coreHelper:
                 return pd.DataFrame(core_for_update).to_string(index=False, justify='left') if core_for_update else "All modules are up-to-date."
             return core_for_update if core_for_update else []
 
-        def install(self):
+        def install(self, dev: bool = False):
             """
-            Method updates helper core.
+            **Method updates helper core.**
+            ```
             Usage:
                 core update install
+            ```
             """
-            core_for_update = self.check(pretty=False)
+            core_for_update = self.check(pretty=False, dev=dev)
             if not core_for_update:
                 return "Nothing to install."
             if core_for_update:
                 core_for_update = core_for_update[0]
                 return install_update(core_for_update['remote_version'], core_for_update['remote_download_link'])
+
+    class logs:
+        """
+        **Module to manipulate with logs.**
+        """
+        def __init__(self, settings: dict):
+            self.settings = settings
+
+        def ls(self, pretty: bool = True):
+            """
+            **Method displays logs.**
+            ```
+            Usage:
+                core logs ls
+            ```
+            """
+            logs = []
+            refactored_logs = []
+            logs.extend(glob.glob(os.path.join(Path(LOGS_PATH), "*.log")))
+            for log in logs:
+                refactored_logs.append(Path(log).stem)
+            return f'Available logs ({len(refactored_logs)}):\n' + '\n'.join(refactored_logs) if pretty else refactored_logs
+
+        def view(self, log: str = None):
+            """
+            **Method displays log content.**
+            ```
+            Usage:
+                core logs view <log>
+            ```
+            """
+            if not log:
+                logs_list = self.ls(pretty=False)
+                log = print_choices(logs_list, exit_btn=True)
+                if log:
+                    log = f"{LOGS_PATH}/{log}.log"
+            else:
+                if not Path(log).exists():
+                    log = f"{LOGS_PATH}/{log}.log"
+                    if not Path(log).exists():
+                        log = None
+            if log:
+                with open(log, 'r', encoding='utf-8') as log_file:
+                    return render_code(log_file.read(), 'python')
+
+
+        def rm(self, log: str = None):
+            """
+            **Method removes log file.**
+            ```
+            Usage:
+                core logs rm
+            ```
+            """
+            if not log:
+                logs_list = self.ls(pretty=False)
+                log = print_choices(logs_list, exit_btn=True)
+                if log:
+                    log = f"{LOGS_PATH}/{log}.log"
+            else:
+                if not Path(log).exists():
+                    log = f"{LOGS_PATH}/{log}.log"
+                    if not Path(log).exists():
+                        log = None
+            if log:
+                os.remove(log)
+                return f"Log file '{Path(log).as_posix()}' removed."
+
+        def flush(self):
+            """
+            **Method removes all log files.**
+            ```
+            Usage:
+                core logs flush
+            ```
+            """
+            logs_list = self.ls(pretty=False)
+            if logs_list:
+                for log in logs_list:
+                    log = f"{LOGS_PATH}/{log}.log"
+                    os.remove(log)
+                return 'All logs removed.'