+++ CHANGELOG.md
@@ -1,4 +1,8 @@
-## 1.1.0 (2026-06-15)
+## 1.1.0 (2026-06-25)
+
+### Global updates
+
+* Markdown rendering support for methods docstrings
 
 ### Modules module
 
@@ -6,7 +10,10 @@
 * Added information about builtin module or not
 * Added push to GitHub function
 * Fixed bug with incorrect hashing and installation when modules has nested folder structure
+* Changed module template
 
 ### Core module
 
 * Fixed bug with update installation
+* Added logs subclass
+* Added `--dev` flag for `core update install` and `core update check`